+++ package.json
@@ -7,7 +7,6 @@
   "scripts": {
     "start": "node index.js",
     "dev": "nodemon index.js",
-    "dev2": "nodemon index2.js",
     "type": "module"
   },
   "repository": {
@@ -21,19 +20,21 @@
   },
   "homepage": "https://github.com/LearnDev-Foundation/learndev-auth-server#readme",
   "dependencies": {
+    "@vonage/server-sdk": "^2.11.2",
     "bcrypt": "^5.0.1",
-    "body-parser": "^1.20.0",
     "connect-mongo": "^4.6.0",
     "cookie-parser": "^1.4.6",
     "cors": "^2.8.5",
     "crypto": "^1.0.1",
     "dayjs": "^1.11.5",
-    "dotenv": "^16.0.2",
+    "dotenv": "^16.0.1",
+    "ejs": "^3.1.8",
     "express": "^4.18.1",
     "express-session": "^1.17.3",
-    "helmet": "^6.0.0",
+    "helmet": "^5.1.1",
     "joi": "^17.6.0",
-    "mongoose": "^6.6.1",
+    "jsonwebtoken": "^8.5.1",
+    "mongoose": "^6.5.2",
     "morgan": "^1.10.0",
     "nodemailer": "^6.7.8",
     "passport": "^0.6.0",
